robot_api/src/infrastructure/messaging/flink_job.py
from pyflink.datastream import StreamExecutionEnvironment
from pyflink.table import StreamTableEnvironment

class FlinkJob:
    def __init__(self):
        self.env = StreamExecutionEnvironment.get_execution_environment()
        self.table_env = StreamTableEnvironment.create(self.env)

    def execute_job(self):
        # Define your Flink job logic here
        # For example, read from a Kafka topic and process the stream
        self.table_env.execute_sql("""
            CREATE TABLE kafka_source (
                id BIGINT,
                name STRING,
                status STRING
            ) WITH (
                'connector' = 'kafka',
                'topic' = 'robot_events',
                'properties.bootstrap.servers' = 'localhost:9092',
                'format' = 'json'
            )
        """)

        self.table_env.execute_sql("""
            CREATE TABLE processed_events (
                id BIGINT,
                name STRING,
                status STRING,
                processed_time TIMESTAMP
            ) WITH (
                'connector' = 'print'
            )
        """)

        self.table_env.execute_sql("""
            INSERT INTO processed_events
            SELECT id, name, status, CURRENT_TIMESTAMP
            FROM kafka_source
        """)

        self.env.execute("Flink Job")

