robot_api/requirements.txt
fastapi
uvicorn
pytest
pylint
pandera
sqlalchemy
pydantic
psycopg2-binary
kafka-python
minio
mlflow
pulumi
transformers
datasets
torch
